Add ability to retrieve table cell info though selection.
The Access Bridge returns incomplete information when accessing
table cells though the regular list of children. The only way to work
around that limitation is to first update the table selection so that
a cell is selected, then retrieve the accessible context of the
selection.

This change introduces a "Select Cells" property group in tables
that does exactly that when expanded. Note that expanding this node
will affect the running java application, in the sense that the
table selection will change as nodes are expanded.