Improve "sync tree" by performing a 3rd pass
If a node can't be found after the first two passes (simple dfs search,
refresh jvm list + simple dfs search), a third pass has been added.
It looks for the top level window owning the node belongs to,
and forces a refreh of that node, so that the children will
be refreshed all the way down to the node.

This is useful when the list of children of a given node changes
over time (for example, a "page tab list" where a new page has just
been activated).